requirements.txt
wyze-sdk
google-genai
